backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.0
networkx>=3.0
pandas==2.2.2
pydantic==2.7.4
python-dateutil==2.9.0
faker==25.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
